requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
jinja2
python-multipart
PyQt6
google-generativeai
google-genai
deepl
requests
